backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.3.4
email-validator==2.1.1